+++ ClevisProjectGroup1/src/hk/edu/polyu/comp/comp2021/clevis/model/Rectangle.java
@@ -3,11 +3,11 @@ package hk.edu.polyu.comp.comp2021.clevis.model;
 /** Shape of Rectangle*/
 class Rectangle implements Shape{
     private final String name; // store name
-    private float w,h; // store the width and height
+    private double w,h; // store the width and height
     private Line[] FourLines = new Line[4]; // LineArray to store 4 line of the rectangle (0: lineLeft, 1: lineBottom, 2: lineRight, 3: lineTop)
 
     /** constructor */
-    Rectangle (String inName, float inX, float inY, float inW, float inH){
+    Rectangle (String inName, double inX, double inY, double inW, double inH){
         name = inName;
         w = inW;
         h = inH;
@@ -17,8 +17,8 @@ class Rectangle implements Shape{
         FourLines[3] = new Line(name+"LineTop", inX+inW,inY,inX,inY); // set the top Line DA
     }
 
-    public float getWidth(){return w;} // get the width
-    public float getHeight(){return h;} // get the height
+    public double getWidth(){return w;} // get the width
+    public double getHeight(){return h;} // get the height
     public Line[] getFourLines(){ return FourLines; } // get 4 Lines of the rectangle
     public Line getWhichFourLines(int i){ // get which Line of the rectangle
         return getFourLines()[i];
@@ -60,17 +60,17 @@ class Rectangle implements Shape{
     }
 
     /** move a Rectangle(and its subclass Square) method*/
-    public void move(float inDx,float inDy){
+    public void move(double inDx,double inDy){
         for (Line i: this.getFourLines()){
             i.move(inDx,inDy);
         }
     }
 
     /** bounding box method */
-    public float getLeftBounding(){return getWhichFourLines(0).getA().getX();}
-    public float getRightBounding(){return getWhichFourLines(2).getA().getX();}
-    public float getTopBounding(){return getWhichFourLines(3).getA().getY();}
-    public float getBottomBounding(){return getWhichFourLines(1).getA().getY();}
+    public double getLeftBounding(){return getWhichFourLines(0).getA().getX();}
+    public double getRightBounding(){return getWhichFourLines(2).getA().getX();}
+    public double getTopBounding(){return getWhichFourLines(3).getA().getY();}
+    public double getBottomBounding(){return getWhichFourLines(1).getA().getY();}
 
     /** list out information of a shape*/
     public String listInfo(){
